+++ src/Listener/SentryCommandListenerFactory.php
@@ -35,6 +35,7 @@ class SentryCommandListenerFactory
             isolateScope: $configReader->requiredBool('sentry_psr.isolate_console_scope'),
             flushOnTerminate: $configReader->requiredBool('sentry_psr.flush_on_console_terminate'),
             captureConsoleInput: $configReader->requiredBool('sentry_psr.capture_console_input'),
+            logConsoleCommandStart: $configReader->requiredBool('sentry_psr.log_console_command_start'),
             logger: LoggerHelper::getLogger($containerResolver),
             sentryLifecycle: $containerResolver->get(SentryLifecycle::class),
             redactor: $redactor,